trading-bot/requirements.txt
ccxt==4.4.27
pandas>=2.3.2
pandas-ta==0.4.71b0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
